refactor(disk): move usage measurement beside the disk primitives
Close the artificial store and worktree upward edges without changing
byte accounting or its hardlink and symlink rules.
