+++ crates/rimz/src/disk/mod.rs
@@ -9,3 +9,4 @@ pub(crate) mod parse_cache;
 pub mod paths;
 pub(crate) mod rotating;
 pub mod single_flight;
+pub mod usage;
